requirements.txt
Pillow>=9.0.0
PyYAML==6.0
schedule~=1.2.0
tweepy~=4.14.0